requirements.txt.txt
customtkinter
pyusb
libusb-package
pyserial
requests
pillow